Move the WebSafe implementation to be the official
